Move pickers to infrastructure js_only
