+++ assets/src/application/ui/forms/espressoForm/adapters/DatePicker.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Datepicker } from '@appInputs/dateTime';
+import { Datepicker } from '@infraUI/inputs/dateTime';
 import { FieldRendererProps } from '../types';
 
 const DatePicker: React.FC<FieldRendererProps> = ({ input: { onChange, ...input }, meta, ...rest }) => {
